+++ test/lib/util/reader-util.test.js
@@ -1,5 +1,56 @@
-const { treatUrl, treatMapping } = require("../../../src/lib/util/reader-util");
-global.console = { error: jest.fn() };
+const {
+  treatUrl,
+  treatMapping,
+  executeUrlExpressions
+} = require("../../../src/lib/util/reader-util");
+const { logger } = require("../../../src/lib/common");
+jest.mock("../../../src/lib/common");
+
+test("executeUrlExpressions without expressions", () => {
+  // Arrange
+  const string =
+    "https://raw.githubusercontent.com/kiegroup/droolsjbpm-build-bootstrap/main/.ci/pull-request-config.yaml";
+
+  // Act
+  const result = executeUrlExpressions(string);
+
+  // Assert
+  expect(result).toBe(string);
+});
+
+test("executeUrlExpressions with 1 expression", () => {
+  // Arrange
+  process.env = Object.assign(process.env, {
+    GITHUB_BASE_REF: "1.2.3"
+  });
+  const string =
+    "https://raw.githubusercontent.com/kiegroup/droolsjbpm-build-bootstrap/%{process.env.GITHUB_BASE_REF.replace(/(\\d*)\\.(.*)\\.(.*)/g, (m, n1, n2, n3) => `${+n1+7}.${n2}.${n3}`)}/.ci/pull-request-config.yml";
+
+  // Act
+  const result = executeUrlExpressions(string);
+
+  // Assert
+  expect(result).toBe(
+    "https://raw.githubusercontent.com/kiegroup/droolsjbpm-build-bootstrap/8.2.3/.ci/pull-request-config.yml"
+  );
+});
+
+test("executeUrlExpressions with 2 expressions", () => {
+  // Arrange
+  process.env = Object.assign(process.env, {
+    GITHUB_BASE_REF: "1.2.3"
+  });
+  const string =
+    "https://raw.githubusercontent.com/%{process.env.GITHUB_BASE_REF.replace(/(\\d*)\\.(.*)\\.(.*)/g, (m, n1, n2, n3) => `${+n1-1}.${+n2+2}.${+n3+3}`)}/droolsjbpm-build-bootstrap/%{process.env.GITHUB_BASE_REF.replace(/(\\d*)\\.(.*)\\.(.*)/g, (m, n1, n2, n3) => `${+n1+7}.${n2}.${n3}`)}/.ci/pull-request-config.yml";
+
+  // Act
+  const result = executeUrlExpressions(string);
+
+  // Assert
+  expect(result).toBe(
+    "https://raw.githubusercontent.com/0.4.6/droolsjbpm-build-bootstrap/8.2.3/.ci/pull-request-config.yml"
+  );
+});
 
 test("treatUrl", () => {
   // Arrange
@@ -49,6 +100,28 @@ test("treatMapping no expression", () => {
 
   // Assert
   expect(mapping).toBe(mapping);
+});
+
+test("treatUrl with expression", () => {
+  // Arrange
+  const string = "1${TWO}${THREE}45%{+20+50}67${EIGHT}";
+
+  // Act
+  const result = treatUrl(string, { TWO: 2, THREE: 3, EIGHT: 8 });
+
+  // Assert
+  expect(result).toBe("1234570678");
+});
+
+test("treatUrl with expression", () => {
+  // Arrange
+  const string = "1${TWO}${THREE}45%{'text'}67${EIGHT}";
+
+  // Act
+  const result = treatUrl(string, { TWO: 2, THREE: 3, EIGHT: 8 });
+
+  // Assert
+  expect(result).toBe("12345text678");
 });
 
 test("treatMapping with simple expression on dependencies", () => {
@@ -333,5 +406,5 @@ test("treatMapping with error expression on dependencies", () => {
 
   // Assert
   expect(mapping).toStrictEqual(expecteResult);
-  expect(console.error).toBeCalled();
+  expect(logger.error).toBeCalled();
 });