+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - javascript expression in definition file URL
 - placeholders replacement with default values
 - [ISSUE 103] be able to access definition file from private repositories
+- [ISSUE 168] dependencies.exclude on project-dependencies tests added
 
 # V2.1.0